fix: moved the logic from the component App to Counter
