+++ my-advanced-profile/src/components/Counter/index.jsx
@@ -1,7 +1,17 @@
-import React from 'react';
+import React, {useState} from 'react';
 import './style.css'
 
-export const Counter = ({count, incrementCount, decrementCount}) => {
+export const Counter = () => {
+
+    const [count, setCount] = useState(0)
+
+    const incrementCount = () => {
+        setCount(count + 1)
+    }
+    const decrementCount = () => {
+        setCount(count - 1)
+    }
+
     return (
         <div className="counter-wrapper">
             { `Cчетчик: ${count < 0 ? "Вы не можете опускаться ниже нуля!" : count}`}
