+++ my-advanced-profile/src/components/ProfileHobbies/index.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import './style.css';
 
+
 function ProfileHobbies({hobbies}) {
     return (
         <>
