+++ my-advanced-profile/src/index.jsx
@@ -6,3 +6,7 @@ import App from "./App";
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(<App />);
+
+
+
+
